translator/requirements.txt
websockets==12.0
aiohttp==3.9.5
azure-cognitiveservices-speech==1.38.0
python-dotenv==1.0.1